fix: memory access violation
the stupid cast would lead to negative numbers, but only for 'obscure'
names, and then negative offset

+++ proteindb.cpp
@@ -138,7 +138,7 @@ void ProteinDB::updateColorset(const QVector<QColor> &colors)
 
 QColor ProteinDB::colorFor(const Protein &subject)
 {
-	return colorset[(int)qHash(subject.name) % colorset.size()];
+	return colorset[(int)(qHash(subject.name) % (unsigned)colorset.size())];
 }
 
 ProteinId ProteinDB::Public::find(const QString &name) const {
